Add utility functions to utils
